Setting up project
