+++ build.gradle
@@ -55,8 +55,8 @@ javafx {
 }
 
 application {
-    mainModule = "com.mycompany.app"
-    mainClass = "com.mycompany.app.App"
+    mainModule = "com.mbcoder.iot.nature_camera"
+    mainClass = "com.mbcoder.iot.nature_camera.NatureCameraApp"
 }
 
 task copyNatives(type: Copy) {
