+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'java-gradle-starter-project'
+rootProject.name = 'nature-camera'
