+++ src/main/java/module-info.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-module com.mycompany.app {
+module com.mbcoder.iot.nature_camera {
   // modules required by the app
-  requires com.esri.arcgisruntime;
+  //requires com.esri.arcgisruntime;
   requires javafx.graphics;
   requires org.slf4j.nop;
 
-  exports com.mycompany.app;
+  exports com.mbcoder.iot.nature_camera;
 }
